server/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2

# Database
psycopg2-binary==2.9.9
django-cleanup==8.0.0
django-phonenumber-field==7.2.1
phonenumbers==8.13.24

# File Storage
Pillow==11.0.0
boto3==1.34.28
django-storages==1.14.2

# API Documentation
drf-yasg==1.21.7
coreapi==2.3.3

# Search and Filtering
django-filter==23.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Email
sendgrid==6.11.0
django-ses==3.5.0

# Caching and Performance
redis==5.0.1
django-redis==5.4.0

# Payments
stripe==7.11.0
django-stripe==6.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
responses==0.24.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
django-stubs==4.2.2
drf-stubs==1.9.0

# Security
python-jose==3.3.0
cryptography==41.0.5

# Utilities
python-slugify==8.0.1
requests==2.31.0
python-magic==0.4.27
django-health-check==3.18.2
